fix review changes
