+++ .gitignore
@@ -47,96 +47,4 @@ __pycache__/
 coverage
 
 yarn-error.log
-generated/backend/AcceptedTosVersion.ts
-generated/backend/BlockedInboxOrChannels.ts
-generated/backend/CieUserIdentity.ts
-generated/backend/CodiceContestoPagamento.ts
-generated/backend/CreatedMessageWithContent.ts
-generated/backend/CreatedMessageWithoutContent.ts
-generated/backend/CreatedMessageWithoutContentCollection.ts
-generated/backend/DepartmentName.ts
-generated/backend/EmailAddress.ts
-generated/backend/EnteBeneficiario.ts
-generated/backend/ExtendedProfile.ts
-generated/backend/FiscalCode.ts
-generated/backend/HasProfile.ts
-generated/backend/Iban.ts
-generated/backend/ImportoEuroCents.ts
-generated/backend/InitializedProfile.ts
-generated/backend/Installation.ts
-generated/backend/InstallationID.ts
-generated/backend/IsEmailEnabled.ts
-generated/backend/IsEmailSet.ts
-generated/backend/IsEmailValidated.ts
-generated/backend/IsInboxEnabled.ts
-generated/backend/IsWebhookEnabled.ts
-generated/backend/LimitedProfile.ts
-generated/backend/MessageBodyMarkdown.ts
-generated/backend/MessageContent.ts
-generated/backend/MessageResponseNotificationStatus.ts
-generated/backend/MessageResponseWithContent.ts
-generated/backend/MessageSubject.ts
-generated/backend/NotificationChannel.ts
-generated/backend/NotificationChannelStatusValue.ts
-generated/backend/OrganizationName.ts
-generated/backend/PageSize.ts
-generated/backend/PaginatedCreatedMessageWithoutContentCollection.ts
-generated/backend/PaginatedServiceTupleCollection.ts
-generated/backend/PaginationResponse.ts
-generated/backend/PaymentActivationsGetResponse.ts
-generated/backend/PaymentActivationsPostRequest.ts
-generated/backend/PaymentActivationsPostResponse.ts
-generated/backend/PaymentAmount.ts
-generated/backend/PaymentData.ts
-generated/backend/PaymentNoticeNumber.ts
-generated/backend/PaymentProblemJson.ts
-generated/backend/PaymentRequestsGetResponse.ts
-generated/backend/Platform.ts
-generated/backend/PreferredLanguage.ts
-generated/backend/PreferredLanguages.ts
-generated/backend/ProblemJson.ts
-generated/backend/Profile.ts
-generated/backend/PublicSession.ts
-generated/backend/PushChannel.ts
-generated/backend/RptId.ts
-generated/backend/ServiceId.ts
-generated/backend/ServiceName.ts
-generated/backend/ServicePublic.ts
-generated/backend/ServiceScope.ts
-generated/backend/ServiceTuple.ts
-generated/backend/SessionInfo.ts
-generated/backend/SessionsList.ts
-generated/backend/SpezzoneStrutturatoCausaleVersamento.ts
-generated/backend/SpezzoniCausaleVersamento.ts
-generated/backend/SpezzoniCausaleVersamentoItem.ts
-generated/backend/SpidLevel.ts
-generated/backend/SpidUserIdentity.ts
-generated/backend/SuccessResponse.ts
-generated/backend/Timestamp.ts
-generated/backend/TimeToLiveSeconds.ts
-generated/backend/UserDataProcessing.ts
-generated/backend/UserDataProcessingChoice.ts
-generated/backend/UserDataProcessingChoiceRequest.ts
-generated/backend/UserDataProcessingStatus.ts
-generated/backend/UserIdentity.ts
-generated/backend/UserMetadata.ts
-generated/backend/Version.ts
-generated/notifications/CreatedMessageWithContent.ts
-generated/notifications/CreatedMessageWithoutContent.ts
-generated/notifications/DepartmentName.ts
-generated/notifications/FiscalCode.ts
-generated/notifications/MessageBodyMarkdown.ts
-generated/notifications/MessageContent.ts
-generated/notifications/MessageSubject.ts
-generated/notifications/Notification.ts
-generated/notifications/NotificationMessage.ts
-generated/notifications/OrganizationName.ts
-generated/notifications/PaymentAmount.ts
-generated/notifications/PaymentData.ts
-generated/notifications/PaymentNoticeNumber.ts
-generated/notifications/ProblemJson.ts
-generated/notifications/SenderMetadata.ts
-generated/notifications/ServiceName.ts
-generated/notifications/SuccessResponse.ts
-generated/notifications/Timestamp.ts
-generated/notifications/TimeToLiveSeconds.ts
+generated/**/*
