+++ HandleNHNotificationCallActivity/function.json
@@ -6,5 +6,5 @@
       "direction": "in"
     }
   ],
-  "scriptFile": "../dist/NHCallServiceActivity/index.js"
+  "scriptFile": "../dist/HandleNHNotificationCallActivity/index.js"
 }
