+++ HandleNHNotificationCallOrchestrator/__tests__/handler.test.ts
@@ -2,26 +2,24 @@
 
 import { context as contextMock } from "../../__mocks__/durable-functions";
 
+import { NonEmptyString } from "italia-ts-commons/lib/strings";
 import { PlatformEnum } from "../../generated/backend/Platform";
-import {
-  FiscalCodeHash,
-  NotificationHubCreateOrUpdateMessage
-} from "../../NHCallService";
+import { NotificationHubCreateOrUpdateMessage } from "../../HandleNHNotificationCall";
 import {
   ActivityInput as NHCallServiceActivityInput,
   ActivityResult
-} from "../../NHCallServiceActivity/handler";
+} from "../../HandleNHNotificationCallActivity/handler";
 import {
   handler,
   OrchestratorInput as NhCallOrchestratorInput
 } from "../handler";
 
-const aFiscalCodeHash = "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855" as FiscalCodeHash;
+const aFiscalCodeHash = "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855" as NonEmptyString;
 const aPushChannel =
   "fLKP3EATnBI:APA91bEy4go681jeSEpLkNqhtIrdPnEKu6Dfi-STtUiEnQn8RwMfBiPGYaqdWrmzJyXIh5Yms4017MYRS9O1LGPZwA4sOLCNIoKl4Fwg7cSeOkliAAtlQ0rVg71Kr5QmQiLlDJyxcq3p";
 const aNotificationHubMessage: NotificationHubCreateOrUpdateMessage = {
   installationId: aFiscalCodeHash,
-  kind: "CreateOrUpdateKind",
+  kind: "CreateOrUpdate",
   platform: PlatformEnum.apns,
   pushChannel: aPushChannel,
   tags: [aFiscalCodeHash]
