+++ HandleNHNotificationCallOrchestrator/function.json
@@ -6,5 +6,5 @@
       "direction": "in"
     }
   ],
-  "scriptFile": "../dist/NHCallOrchestrator/index.js"
+  "scriptFile": "../dist/HandleNHNotificationCallsOrchestrator/index.js"
 }
