+++ package.json
@@ -12,9 +12,9 @@
     "test": "jest",
     "test:coverage": "jest --coverage",
     "lint": "tslint -p .",
-    "generate:proxy-models": "npm-run-all generate:proxy:*",
-    "generate:proxy:api-models": "rimraf generated/backend && shx mkdir -p generated/backend && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-backend/master/api_backend.yaml --out-dir generated/backend",
-    "generate:proxy:notification-models": "rimraf generated/notifications && shx mkdir -p generated/notifications && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-backend/master/api_notifications.yaml --out-dir generated/notifications",
+    "generate:backend-models": "npm-run-all generate:backend:*",
+    "generate:backend:api-models": "rimraf generated/backend && shx mkdir -p generated/backend && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-backend/master/api_backend.yaml --out-dir generated/backend",
+    "generate:backend:notification-models": "rimraf generated/notifications && shx mkdir -p generated/notifications && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-backend/master/api_notifications.yaml --out-dir generated/notifications",
     "deploy": "npm run build && func azure functionapp publish agid-functions-app-test",
     "dist:modules": "modclean -r -n default:safe && yarn install --production",
     "predeploy": "npm-run-all build dist:modules"
